Resolve dependencies for generator-created cores
Currently, generated cores (i.e. core files written by a `generator`)
are inserted into the list of cores, but their dependencies are ignored.
After this commit, cores are inserted into the list of available cores,
and their dependencies are evaluated as with normal cores.

The implementation is actually simple:

- Generators produce files, including .core files, in a special
  directory, unique per generator run.
- This directory is added as library to the core database.
- All cores in this new library (i.e. all .core files) are added as
  dependency to the current project.

With that in place, the normal dependency resolution process takes over.

This new behavior is actually easier to understand for users, as
generator-produced are no longer a special case, but handled exactly as
other cores.

NOTE: This commit also includes another behavior change: the
`generate.*.position` attribute is no longer supported, files are always
inserted according to the normal dependency resolution orders. The test
is changed accordingly. This needs to be documented better and a
deprecation warning issued.

Fixes #365