Improve handling of generated cores
With this commit, is_generated is set for all generated cores. This is
needed to later on relocate generator outputs and to delete the ttptttg
temporary working directories afterwards.

Without this commit, the temporary working directories are not deleted
causing intermediate results from many fusesoc runs to pile up in /tmp.
In addition, when using the --no-export argument, no generator outputs
are relocated out of /tmp which breaks some tool flows.